requirements.txt
flask
pillow
google-generativeai
gunicorn